@@ -12,6 +12,7 @@
     "build:proto": "pwsh Generate-Proto.ps1",
     "build:typescript": "tsc",
     "build:copy": "npx copyfiles -u 1 ./src/proto/**/*_pb.* ./lib/",
+    "test:typescript": "jest",
     "test:build": "npx webpack build",
     "test:jasmine": "npx jasmine-browser-runner runSpecs",
     "test:debug": "npx jasmine-browser-runner serve",
@@ -35,31 +36,36 @@
   ],
   "license": "ISC",
   "dependencies": {
+    "@improbable-eng/grpc-web-node-http-transport": "^0.15.0",
     "@trinsic/okapi": "^1.4.0",
-    "google-protobuf": "^3.17.2",
-    "js-base64": "^3.7.2",
-    "nice-grpc-web": "^1.0.8",
+    "base64url": "^3.0.1",
     "buffer": "^6.0.3",
-    "base64url": "^3.0.1"
+    "google-protobuf": "^3.17.2",
+    "grpc-web-node-http-transport": "0.0.1",
+    "js-base64": "^3.7.2",
+    "nice-grpc-web": "^1.0.8"
   },
   "devDependencies": {
     "@types/google-protobuf": "^3.15.5",
+    "@types/jest": "^27.4.1",
     "@types/node": "^15.12.0",
     "chromedriver": "^100.0.0",
     "copyfiles": "^2.4.1",
     "dotenv": "^10.0.0",
     "geckodriver": "^2.0.0",
+    "jasmine": "^4.1.0",
     "jasmine-browser-runner": "^1.0.0",
-    "jasmine-core": "^4.0.0",
-    "jest": "^27.0.4",
+    "jasmine-core": "^4.1.0",
+    "jest": "^27.5.1",
     "karma": "^6.3.3",
     "karma-chrome-launcher": "^3.1.0",
     "karma-firefox-launcher": "^2.1.1",
     "karma-jasmine": "^4.0.1",
     "karma-safari-launcher": "^1.0.0",
+    "ts-jest": "^27.1.4",
     "ts-node": "^10.0.0",
     "ts-proto": "^1.110.4",
-    "typescript": "^4.3.2",
+    "typescript": "^4.6.3",
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2"
   }
