@@ -1,4 +1,4 @@
-const { ServiceOptions } = require("../lib/proto");
+import { ServiceOptions } from "../src"
 
 export const options = ServiceOptions.fromPartial({serverEndpoint: "staging-internal.trinsic.cloud",
     serverPort: 443, serverUseTls: true})
